ELF: Allow forward references to linked sections.
It's possible to create IR that uses !associated to refer to a global that
appears later in the module, which can result in these types of forward
references being generated. Unfortunately our assembler does not currently
accept the resulting .s so I needed to use yaml2obj to test this.

Differential Revision: https://reviews.llvm.org/D64880

git-svn-id: http://llvm.org/svn/llvm-project/lld/trunk@366460 91177308-0d34-0410-b5e6-96231b3b80d8